made encoding files in unicode
